test/fixtures/wpt/FileAPI/blob/Blob-slice-overflow.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Blob slice overflow</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/FileAPI/#dfn-slice">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

var text = '';

for (var i = 0; i < 2000; ++i) {
  text += 'A';
}

test(function() {
  var blob = new Blob([text]);
  var sliceBlob = blob.slice(-1, blob.size);
  assert_equals(sliceBlob.size, 1, "Blob slice size");
}, "slice start is negative, relativeStart will be max((size + start), 0)");

test(function() {
  var blob = new Blob([text]);
  var sliceBlob = blob.slice(blob.size + 1, blob.size);
  assert_equals(sliceBlob.size, 0, "Blob slice size");
}, "slice start is greater than blob size, relativeStart will be min(start, size)");

test(function() {
  var blob = new Blob([text]);
  var sliceBlob = blob.slice(blob.size - 2, -1);
  assert_equals(sliceBlob.size, 1, "Blob slice size");
}, "slice end is negative, relativeEnd will be max((size + end), 0)");

test(function() {
  var blob = new Blob([text]);
  var sliceBlob = blob.slice(blob.size - 2, blob.size + 999);
  assert_equals(sliceBlob.size, 2, "Blob slice size");
}, "slice end is greater than blob size, relativeEnd will be min(end, size)");

</script>
